+++ CHANGELOG.md
@@ -7,11 +7,13 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 
 ## [Unreleased]
 
+
 ### Breaking changes
 
-* All `HashMap<&str, MetadataValue<D>>` have been replaced with `metadata::Map<D>` (and they no longer have a lifetime associated)
+* All `HashMap<&str, MetadataValue<D>>` have been replaced with `metadata::Map<D>`
 * The `Posting` no longer have a lifetime
 * The fields `payee` and `narration` in `Transaction` are now of type `Option<String>`
+* The fields `tags` and `links` in `Transaction` are now of type `HashSet<Tag>` and `HashSet<Link>` respectively
 
 
 ### Deprecated